% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Route.R
\name{Route}
\alias{Route}
\title{Route class}
\description{
This class represents assigning an R function to an http request path
}
\examples{
Route$new(path="/flight_column_names", handler = getColumnNames, content_type = "text/plain")

}
\author{
Rick Dean
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Is a string that sets the http request path.}

\item{\code{handler}}{Is the \code{Route} function handler which optionally accepts
a \code{\link{Request}} argument.}

\item{\code{content_type}}{Is the MIME content type returned by the function. Common MIME types are
listed \href{https://developer.mozilla.org/en-US/docs/Web/HTTP/Basics_of_HTTP/MIME_types}{here}.
The default MIME type is "text/plain".}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Route-new}{\code{Route$new()}}
\item \href{#method-Route-clone}{\code{Route$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Route-new"></a>}}
\if{latex}{\out{\hypertarget{method-Route-new}{}}}
\subsection{Method \code{new()}}{
Create a Route object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Route$new(path = NULL, handler = NULL, content_type = "text/plain")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{http request path.}

\item{\code{handler}}{A callback function that executes when the \code{path} request is received.}

\item{\code{content_type}}{A string that defines the MIME type of the content returned by \code{handler}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Route} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Route-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Route-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Route$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
