% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server.R
\name{Server}
\alias{Server}
\title{Server class}
\description{
The package is based on \href{https://cran.r-project.org/web/packages/httpuv/index.html}{httpuv::}.
It is designed to host static resources such as html, javascript, css, etc along with specific route requests
to perform various R based actions or respond with various content(i.g. JSON strings, images, etc).  The server
is designed to work in a local host environment (i.e. url = 127.0.0.1) as part of an html/javascript project that
needs a link to R language capabilities and its associated packages.
}
\details{
Class provides an http server that connects R language resources with javascript applications.
}
\author{
Rick Dean
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{index_path}}{A string that defines the full directory path to an \code{index.html} file defining the server's
root path (i.e. "\"). The root html file is assumed to be named "index.html".}

\item{\code{port}}{An integer that defines the port number for the server. The default is 8080.}

\item{\code{routes}}{A vector of \code{RserverPkg::Route} objects that define the request paths that the server will respond to.
Each \code{RserverPkg::Route} object should define the path, callback function that optionally returns content, and the
content's MIME content type. The callback function can be defined to receive a \code{RserverPkg::Request} object argument.}

\item{\code{static_paths}}{A named list that defines the locations of static resources (.css, .html, .js files) for hosting by the
server. The list name is the web path (e.g. "/content") and the value is the full directory path to the resources
(e.g. F:/web_server/content).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Server$new()}}
\item \href{#method-start}{\code{Server$start()}}
\item \href{#method-stop}{\code{Server$stop()}}
\item \href{#method-list_servers}{\code{Server$list_servers()}}
\item \href{#method-clone}{\code{Server$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Defines the classes' public fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Server$new(index_path = NULL, port = 8080, routes = NULL, static_paths = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index_path}}{A string that defines the full directory path to an \code{index.html} file defining the server's
root path (i.e. "\").}

\item{\code{port}}{An integer that defines the port number for the server.}

\item{\code{routes}}{A vector of \code{RserverPkg::Route} objects that define the request paths that the server will respond to.}

\item{\code{static_paths}}{A named list that defines the locations of static resources (.css, .html, .js files) for hosting by the
server.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start"></a>}}
\if{latex}{\out{\hypertarget{method-start}{}}}
\subsection{Method \code{start()}}{
Start the server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Server$start()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stop"></a>}}
\if{latex}{\out{\hypertarget{method-stop}{}}}
\subsection{Method \code{stop()}}{
Stop the server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Server$stop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_servers"></a>}}
\if{latex}{\out{\hypertarget{method-list_servers}{}}}
\subsection{Method \code{list_servers()}}{
List the current servers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Server$list_servers()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Server$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
